improve missing cassettes error message

+++ features/support/hooks.rb
@@ -31,6 +31,9 @@ end
 
 Around do |scenario, block|
   VCR.use_cassette(scenario.location.file.chomp('.feature') + "/" + scenario.name.gsub(/[^A-Za-z0-9]+/, '-')[0..100], :record_on_error => false, :match_requests_on => [:method, :host, :path, :query, :body_as_json]) do |cassette|
+    if !File.exist?(cassette.file) && ENV["RECORD"] == "false"
+      raise Exception.new "Cassette '#{cassette.file}' not found: create one setting `RECORD=true` or ignore it using `RECORD=none`";
+    end
     File.delete(cassette.file) if ENV["RECORD"] == "true" && File.exist?(cassette.file) && !scenario.match_tags?("@replay-only")
     Timecop.freeze(use_real_time? ? Time.now : cassette.originally_recorded_at) do
       block.call
